delete redundant script/make annotations
