+++ scripts/01-denue_to_nodes.py
@@ -11,39 +11,24 @@ if module_path not in sys.path:
     import aup
 
 def main(schema, year, save=False):
-    df = pd.read_json("/home/jovyan/work/scripts/metropolis_test.json")
+    df = pd.read_json("/home/jovyan/work/scripts/Metropolis_CVE.json")
     aup.log("Read metropolitan areas and capitals json")
 
     #Folder names from database
-    mpos_folder = 'mpos_'+ year
     denue_folder = 'denue_' + year
 
-    # Iterate over municipality DataFrame columns to access each municipality code
+    # Iterate over cities and download municipalities gdf
     for c in df.columns.unique():
         aup.log(f"\n Starting municipality filters for {c}")
         # Creates empty GeoDataFrame to store specified municipality polygons
         mun_gdf = gpd.GeoDataFrame()
         # Iterates over municipality codes for each metropolitan area or capital
-        for i in range(len(df.loc["mpos", c])):
-            # Extracts specific municipality code
-            m = df.loc["mpos", c][i]
-            # Downloads municipality polygon according to code
-            query = f"SELECT * FROM marco.{mpos_folder} WHERE \"CVEGEO\" LIKE \'{m}\'"
-            mun_gdf = mun_gdf.append(aup.gdf_from_query(query, geometry_col='geometry'))
-            aup.log(f"Downloaded {m} GeoDataFrame at: {c}")    
+        # Downloads municipality polygon according to code
+        query = f"SELECT * FROM metropolis.metro_list WHERE \"city\" LIKE \'{c}\'"
+        mun_gdf = aup.gdf_from_query(query, geometry_col='geometry')
+        aup.log(f"Downloaded municipality GeoDataFrames at: {c}")    
         #Define projections
         mun_gdf = mun_gdf.set_crs("EPSG:4326")
-        
-        #poly = mun_gdf.geometry
-        ## Extracts coordinates from polygon as DataFrame
-        #coord_val = poly.bounds
-        ## Gets coordinates for bounding box
-        #n = coord_val.maxy.max()
-        #s = coord_val.miny.min()
-        #e = coord_val.maxx.max()
-        #w = coord_val.minx.min()
-    # Downloads OSMnx graph from bounding box
-        #G = ox.graph_from_bbox(n, s, e, w, network_type="all")
     
         #Creates polygon for query
         poly_wkt = mun_gdf.dissolve().geometry.to_wkt()[0]
@@ -54,6 +39,7 @@ def main(schema, year, save=False):
         denue = aup.gdf_from_query(query, geometry_col='geometry')
         aup.log(f"Downloaded {len(denue)} denue from database for {c}")
 
+        #Downloads street network from base
         G, nodes, edges = aup.graph_from_hippo(mun_gdf, 'osmnx')
         #Defines projection for downloaded data
         denue = denue.set_crs("EPSG:4326")
@@ -62,12 +48,13 @@ def main(schema, year, save=False):
 
 
         aup.log(f"Created NetworkX for {c}")
-
+        ### Filters relevant columns from DENUE.
+        ### If you are interested in keeping other columns, change/add here
         points = denue[['id', 'codigo_act', 'geometry']]
-
+        ### Calculate nearest node for each DENUE point
         nearest = aup.find_nearest(G, nodes, points, return_distance= True)
         nearest = nearest.set_crs("EPSG:4326")
-
+        ###SAVE
         if save:
             aup.gdf_to_db_slow(nearest, "denue_node_"+year, schema=schema, if_exists="append")
