+++ scripts/02-distance_amenities.py
@@ -16,26 +16,19 @@ def main(schema, folder_sufix, year, amenities, resolution=8, save=False):
     aup.log("Read metropolitan areas and capitals json")
     
     #Folder names from database
-    mpos_folder = 'mpos_'+year
+    denue_folder = 'denue_node_'+year
     
-    # Iterate over municipality DataFrame columns to access each municipality code
+    # Iterate over cities and download municipalities gdf
     for c in df.columns.unique():
         aup.log(f"\n Starting municipality filters for {c}")
         # Creates empty GeoDataFrame to store specified municipality polygons and hex grid
         mun_gdf = gpd.GeoDataFrame()
         hex_bins = gpd.GeoDataFrame()
-        # Iterates over municipality codes for each metropolitan area or capital
-        for i in range(len(df.loc["mpos", c])):
-            # Extracts specific municipality code
-            m = df.loc["mpos", c][i]
-            # Downloads municipality polygon according to code
-            query = f"SELECT * FROM marco.{mpos_folder} WHERE \"CVEGEO\" LIKE \'{m}\'"
-            mun_gdf = mun_gdf.append(aup.gdf_from_query(query, geometry_col='geometry'))
-            aup.log(f"Downloaded {m} GeoDataFrame at: {c}")
-            #Creates query to download hex bins according to code
-            query = f"SELECT * FROM hexgrid.hexgrid_mx WHERE \"CVEGEO\" LIKE \'{m}%%\'"
-            hex_bins = hex_bins.append(aup.gdf_from_query(query, geometry_col='geometry'))
-            aup.log(f"Donwloaded hex bins for {m}")
+        # Iterates over city names for each metropolitan area or capital
+        query = f"SELECT * FROM metropolis.metro_list WHERE \"city\" LIKE \'{c}\'"
+        mun_gdf = aup.gdf_from_query(query, geometry_col='geometry')
+        query = f"SELECT * FROM metropolis.hexgrid_{resolution}_city WHERE \"metropolis\" LIKE \'{c}\'"
+        hex_bins = aup.gdf_from_query(query, geometry_col='geometry')
             
         #Define projections for municipalities and hexgrids
         mun_gdf = mun_gdf.set_crs("EPSG:4326")
@@ -68,7 +61,7 @@ def main(schema, folder_sufix, year, amenities, resolution=8, save=False):
             denue_amenity = gpd.GeoDataFrame()
             #Based on the SCIAN code, the POIs will be downloaded from the DB
             for cod in amenities[a]:
-                query = f"SELECT * FROM denue_nodes.denue_node_2020 WHERE (ST_Intersects(geometry, \'SRID=4326;{poly_wkt}\')) AND (\"codigo_act\" = {cod})"
+                query = f"SELECT * FROM denue_nodes.{denue_folder} WHERE (ST_Intersects(geometry, \'SRID=4326;{poly_wkt}\')) AND (\"codigo_act\" = {cod})"
                 denue_amenity = denue_amenity.append(aup.gdf_from_query(query, geometry_col='geometry'))
             aup.log(f"Downloaded accumulated total of {len(denue_amenity)} {a} from database for {c}")
             df_temp = nodes.copy()
@@ -80,8 +73,9 @@ def main(schema, folder_sufix, year, amenities, resolution=8, save=False):
             if len(denue_amenity) == 0:
                 nodes_distance['time_'+a] = 0
                 aup.log(f"0 {cod} found in {c}")
+                ##### This hecks if the number of points is an exact multiple of 250, if it is it will run with segments of 200, in order to avoid a crash.
             elif len(denue_amenity) % 250:
-                #Due to memory constraints, the total number of POIs will be divided in groups of 100
+                #Due to memory constraints, the total number of POIs will be divided in groups of 250
                 #These will run with the calculate nearest distance poi function by group and will be stored
                 # to check later
                 c_denue = len(denue_amenity)/200
@@ -105,7 +99,7 @@ def main(schema, folder_sufix, year, amenities, resolution=8, save=False):
                 #all final data
                 nodes_distance = nodes_distance.merge(df_min, left_index=True, right_index=True)
             else:
-                #Due to memory constraints, the total number of POIs will be divided in groups of 100
+                #Due to memory constraints, the total number of POIs will be divided in groups of 250
                 #These will run with the calculate nearest distance poi function by group and will be stored
                 # to check later
                 c_denue = len(denue_amenity)/250
